check if preferenxes.xml exist before parsing it

+++ container/ubooquity/build/docker-entrypoint.sh
@@ -20,10 +20,12 @@ else
 fi
 
 # now parse the nginx configuration file with j2
-echo "Parse the configuration file with j2"
-mv "$CONFIG_LOCAL" "$CONFIG_LOCAL.orig"
-j2 "$CONFIG_LOCAL.orig" > "$CONFIG_LOCAL"
-[ $? -ne 0 ] && exit 1
+if [ -f $CONFIG_LOCAL ]; then
+  echo "Parse the configuration file with j2"
+  mv "$CONFIG_LOCAL" "$CONFIG_LOCAL.orig"
+  j2 "$CONFIG_LOCAL.orig" > "$CONFIG_LOCAL"
+  [ $? -ne 0 ] && exit 1
+fi
 
 # run ubooquity
 echo "Run ubooquity"
